feat: add pop-up component

+++ src/app/core/components/header/header.component.ts
@@ -3,10 +3,11 @@ import { AuthService } from 'src/app/auth/services/auth.service';
 
 import { TranslateService } from '@ngx-translate/core';
 
-
 import { MatDialog } from '@angular/material/dialog';
 import { ProfileComponent } from 'src/app/main/components/profile/profile.component';
 import { Router } from '@angular/router';
+import { MatDialogComponent } from '../mat-dialog/mat-dialog.component';
+
 @Component({
   selector: 'app-header',
   templateUrl: './header.component.html',
@@ -24,7 +25,8 @@ export class HeaderComponent {
 
     public translate: TranslateService,
     public prof: MatDialog,
-    private router: Router
+    private router: Router,
+    public dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -36,10 +38,21 @@ export class HeaderComponent {
   }
 
   logOut(): void {
-    // this.isSingIn$.next(false);
-    localStorage.clear();
-    this.router.navigateByUrl('/welcome');
-    this.auth.logOut();
+    const dialogRef = this.dialog.open(MatDialogComponent, {
+      data: {
+        message:
+          localStorage.getItem('lang') === 'ukr'
+            ? 'Ви дісно бажаєте вийти?'
+            : 'Are you sure you want to logout?',
+      },
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.auth.logOut();
+        localStorage.clear();
+        this.router.navigateByUrl('/welcome');
+      }
+    });
   }
 
   saveToLocal(lang: string): void {
